+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Support for Burglar Alarm (WHO=5)
+- checks to Where base class constructor
 
 
 ## [0.7.1] - 2022-01-14